created a separate file for knapsack functions
